add test for bigint primary key and foreign key

+++ src/java/com/rapleaf/jack/test_project/database_1/mock_impl/BaseMockCommentPersistenceImpl.java
@@ -41,14 +41,14 @@ public class BaseMockCommentPersistenceImpl extends AbstractMockDatabaseModel<Co
   public ModelWithId create(Map<Enum, Object> fieldsMap) throws IOException {
     String content = (String) fieldsMap.get(Comment._Fields.content);
     Integer commenter_id = (Integer) fieldsMap.get(Comment._Fields.commenter_id);
-    Integer commented_on_id = (Integer) fieldsMap.get(Comment._Fields.commented_on_id);
+    Long commented_on_id = (Long) fieldsMap.get(Comment._Fields.commented_on_id);
     Long created_at_tmp = (Long) fieldsMap.get(Comment._Fields.created_at);
     long created_at = created_at_tmp == null ? 28800000 : created_at_tmp;
     return create(content, commenter_id, commented_on_id, created_at);
   }
 
 
-  public Comment create(final String content, final Integer commenter_id, final Integer commented_on_id, final long created_at) throws IOException {
+  public Comment create(final String content, final Integer commenter_id, final Long commented_on_id, final long created_at) throws IOException {
     long __id = curId.getAndIncrement();
     Comment newInst = new Comment(__id, content, commenter_id, commented_on_id, created_at, databases);
     records.put(__id, newInst);
@@ -83,7 +83,7 @@ public class BaseMockCommentPersistenceImpl extends AbstractMockDatabaseModel<Co
     return find(new HashMap<Enum, Object>(){{put(Comment._Fields.commenter_id, value);}});
   }
 
-  public Set<Comment> findByCommentedOnId(final Integer value) throws IOException {
+  public Set<Comment> findByCommentedOnId(final Long value) throws IOException {
     return find(new HashMap<Enum, Object>(){{put(Comment._Fields.commented_on_id, value);}});
   }
